Release 1.0.0-beta.3

+++ app/tabs/codeListTab.js
@@ -17,7 +17,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { withStyles } from '@material-ui/core/styles';
 import { shell } from 'electron';
-import {BootstrapTable, ButtonGroup} from 'react-bootstrap-table';
+import { BootstrapTable, ButtonGroup } from 'react-bootstrap-table';
 import deepEqual from 'fast-deep-equal';
 import clone from 'clone';
 import renderColumns from 'utils/renderColumns.js';
@@ -67,27 +67,28 @@ const styles = theme => ({
 // Redux functions
 const mapDispatchToProps = dispatch => {
     return {
-        updateCodeList          : (oid, updateObj) => dispatch(updateCodeList(oid, updateObj)),
-        updateCodeListStandard  : (oid, updateObj) => dispatch(updateCodeListStandard(oid, updateObj)),
-        openModal               : (updateObj) => dispatch(openModal(updateObj)),
-        updateExternalCodeList  : (oid, updateObj) => dispatch(updateExternalCodeList(oid, updateObj)),
-        deleteCodeLists         : (deleteObj) => dispatch(deleteCodeLists(deleteObj)),
+        updateCodeList: (oid, updateObj) => dispatch(updateCodeList(oid, updateObj)),
+        updateCodeListStandard: (oid, updateObj) => dispatch(updateCodeListStandard(oid, updateObj)),
+        openModal: (updateObj) => dispatch(openModal(updateObj)),
+        updateExternalCodeList: (oid, updateObj) => dispatch(updateExternalCodeList(oid, updateObj)),
+        deleteCodeLists: (deleteObj) => dispatch(deleteCodeLists(deleteObj)),
     };
 };
 
 const mapStateToProps = state => {
     return {
-        mdv           : state.present.odm.study.metaDataVersion,
-        codeLists     : state.present.odm.study.metaDataVersion.codeLists,
-        codeListOrder : state.present.odm.study.metaDataVersion.order.codeListOrder,
-        standards     : state.present.odm.study.metaDataVersion.standards,
-        stdCodeLists  : state.present.stdCodeLists,
-        stdConstants  : state.present.stdConstants,
-        defineVersion : state.present.odm.study.metaDataVersion.defineVersion,
-        tabs          : state.present.ui.tabs,
-        tabSettings   : state.present.ui.tabs.settings[state.present.ui.tabs.currentTab],
-        showRowSelect : state.present.ui.tabs.settings[state.present.ui.tabs.currentTab].rowSelect['overall'],
-        reviewMode    : state.present.ui.main.reviewMode,
+        mdv: state.present.odm.study.metaDataVersion,
+        codeLists: state.present.odm.study.metaDataVersion.codeLists,
+        codeListOrder: state.present.odm.study.metaDataVersion.order.codeListOrder,
+        standards: state.present.odm.study.metaDataVersion.standards,
+        stdCodeLists: state.present.stdCodeLists,
+        stdConstants: state.present.stdConstants,
+        defineVersion: state.present.odm.study.metaDataVersion.defineVersion,
+        tabs: state.present.ui.tabs,
+        tabSettings: state.present.ui.tabs.settings[state.present.ui.tabs.currentTab],
+        showRowSelect: state.present.ui.tabs.settings[state.present.ui.tabs.currentTab].rowSelect['overall'],
+        reviewMode: state.present.ui.main.reviewMode,
+        showDeleteCodeListWarning: state.present.settings.popUp.onCodeListDelete,
     };
 };
 
@@ -128,11 +129,11 @@ function openLink (event) {
 // Formatter functions
 function codeListStandardFormatter (cell, row) {
     if (row.codeListType === 'external') {
-        let result='';
+        let result = '';
         if (cell.href === undefined) {
-            result = <div>{(cell.dictionary||'') + ' Version:' + (cell.version||'')}</div>;
+            result = <div>{(cell.dictionary || '') + ' Version:' + (cell.version || '')}</div>;
         } else if (cell.href !== undefined) {
-            result = <a onClick={openLink} href={cell.href}>{(cell.dictionary||'') + ' Version:' + (cell.version||'')}</a>;
+            result = <a onClick={openLink} href={cell.href}>{(cell.dictionary || '') + ' Version:' + (cell.version || '')}</a>;
         }
         if (cell.ref !== undefined) {
             result = <div>{result}<span> Ref: {cell.ref}</span></div>;
@@ -146,7 +147,7 @@ function codeListStandardFormatter (cell, row) {
 function codeListTypeFormatter (cell, row) {
     if (cell !== undefined) {
         let typeDecode;
-        row.stdConstants.codeListTypes.some( type => {
+        row.stdConstants.codeListTypes.some(type => {
             if (type.hasOwnProperty(cell)) {
                 typeDecode = type[cell];
                 return true;
@@ -158,7 +159,7 @@ function codeListTypeFormatter (cell, row) {
 }
 
 class ConnectedCodeListTable extends React.Component {
-    constructor(props) {
+    constructor (props) {
         super(props);
 
         let columns = clone(props.stdConstants.columns.codeLists);
@@ -173,49 +174,49 @@ class ConnectedCodeListTable extends React.Component {
                 dataFormat: this.menuFormatter,
             },
             name: {
-                customEditor: {getElement: simpleInputEditor, customEditorParameters: { options:
+                customEditor: { getElement: simpleInputEditor,
+                    customEditorParameters: { options:
                     {
-                        checkForSpecialChars : { type: 'Note' },
+                        checkForSpecialChars: { type: 'Note' },
                     }
-                }},
+                    } },
             },
             codeListType: {
-                dataFormat   : codeListTypeFormatter,
-                customEditor : {getElement: codeListTypeSelectEditor, customEditorParameters: {options: props.stdConstants.codeListTypes}},
+                dataFormat: codeListTypeFormatter,
+                customEditor: { getElement: codeListTypeSelectEditor, customEditorParameters: { options: props.stdConstants.codeListTypes } },
             },
             dataType: {
-                customEditor: {getElement: simpleSelectEditor, customEditorParameters: {options: props.stdConstants.dataTypes}},
+                customEditor: { getElement: simpleSelectEditor, customEditorParameters: { options: props.stdConstants.dataTypes } },
             },
             formatName: {
-                customEditor: {getElement: codeListFormatNameEditor},
+                customEditor: { getElement: codeListFormatNameEditor },
             },
             linkedCodeList: {
-                customEditor: {getElement: linkedCodeListEditor},
+                customEditor: { getElement: linkedCodeListEditor },
             },
             standardData: {
-                dataFormat   : codeListStandardFormatter,
-                customEditor : {getElement: codeListStandardEditor},
+                dataFormat: codeListStandardFormatter,
+                customEditor: { getElement: codeListStandardEditor },
             },
         };
 
         // Unite Columns with Editors and Formatters;
-        Object.keys(columns).forEach( id => {
+        Object.keys(columns).forEach(id => {
             columns[id] = { ...columns[id], ...editorFormatters[id] };
         });
 
         this.state = {
             columns,
-            anchorEl           : null,
-            selectedRows       : [],
-            codeListMenuParams : {},
-            showAddCodeList    : false,
-            insertPosition     : null,
-            showSelectColumn   : false,
+            anchorEl: null,
+            selectedRows: [],
+            codeListMenuParams: {},
+            showAddCodeList: false,
+            insertPosition: null,
+            showSelectColumn: false,
         };
     }
 
-    static getDerivedStateFromProps(nextProps, prevState) {
-
+    static getDerivedStateFromProps (nextProps, prevState) {
         let columns = getColumnHiddenStatus(prevState.columns, nextProps.tabSettings.columns, nextProps.showRowSelect);
 
         if (!deepEqual(columns, prevState.columns)) {
@@ -224,7 +225,7 @@ class ConnectedCodeListTable extends React.Component {
         return null;
     }
 
-    componentDidMount() {
+    componentDidMount () {
         setScrollPosition(this.props.tabs);
     }
 
@@ -263,17 +264,17 @@ class ConnectedCodeListTable extends React.Component {
                 if (cellValue !== undefined) {
                     let codeList = this.props.codeLists[row.oid];
                     let linkedCodeList = this.props.codeLists[cellValue];
-                    if (codeList.codeListType === 'enumerated'
-                        && codeList.standardOid !== undefined
-                        && this.props.stdCodeLists.hasOwnProperty(codeList.standardOid)
-                        && this.props.stdCodeLists[codeList.standardOid].nciCodeOids.hasOwnProperty(codeList.alias.name)
+                    if (codeList.codeListType === 'enumerated' &&
+                        codeList.standardOid !== undefined &&
+                        this.props.stdCodeLists.hasOwnProperty(codeList.standardOid) &&
+                        this.props.stdCodeLists[codeList.standardOid].nciCodeOids.hasOwnProperty(codeList.alias.name)
                     ) {
                         let standardCodeListOid = this.props.stdCodeLists[codeList.standardOid].nciCodeOids[codeList.alias.name];
                         updateObj.standardCodeList = this.props.stdCodeLists[codeList.standardOid].codeLists[standardCodeListOid];
-                    } else if (linkedCodeList.codeListType === 'enumerated'
-                        && linkedCodeList.standardOid !== undefined
-                        && this.props.stdCodeLists.hasOwnProperty(linkedCodeList.standardOid)
-                        && this.props.stdCodeLists[linkedCodeList.standardOid].nciCodeOids.hasOwnProperty(linkedCodeList.alias.name)
+                    } else if (linkedCodeList.codeListType === 'enumerated' &&
+                        linkedCodeList.standardOid !== undefined &&
+                        this.props.stdCodeLists.hasOwnProperty(linkedCodeList.standardOid) &&
+                        this.props.stdCodeLists[linkedCodeList.standardOid].nciCodeOids.hasOwnProperty(linkedCodeList.alias.name)
                     ) {
                         let standardCodeListOid = this.props.stdCodeLists[linkedCodeList.standardOid].nciCodeOids[linkedCodeList.alias.name];
                         updateObj.standardCodeList = this.props.stdCodeLists[linkedCodeList.standardOid].codeLists[standardCodeListOid];
@@ -297,13 +298,12 @@ class ConnectedCodeListTable extends React.Component {
                 updateObj[cellName] = cellValue;
                 this.props.updateCodeList(row.oid, updateObj);
             }
-
         }
         return true;
     }
 
     cleanSelection = () => {
-        if (this.state.selectedRows.length > 0)  {
+        if (this.state.selectedRows.length > 0) {
             this.setState({ selectedRows: [] });
         }
     }
@@ -373,15 +373,15 @@ class ConnectedCodeListTable extends React.Component {
     createCustomToolBar = props => {
         return (
             <Grid container spacing={16} justify='space-between'>
-                <Grid item style={{paddingLeft: '8px'}}>
+                <Grid item style={{ paddingLeft: '8px' }}>
                     { props.components.btnGroup }
                 </Grid>
-                <Grid item style={{paddingRight: '25px'}}>
+                <Grid item style={{ paddingRight: '25px' }}>
                     <Grid container spacing={16} justify='flex-end'>
                         <Grid item>
                             <Button variant="contained" color="default" onClick={ () => { this.setState({ showSelectColumn: true }); } }>
                                 Columns
-                                <RemoveRedEyeIcon style={{marginLeft: '7px'}}/>
+                                <RemoveRedEyeIcon style={{ marginLeft: '7px' }}/>
                             </Button>
                         </Grid>
                     </Grid>
@@ -410,7 +410,7 @@ class ConnectedCodeListTable extends React.Component {
         // Get the list of ItemOIDs for which the codelists should be removed;
         let itemDefOids = [];
         codeListOids.forEach(codeListOid => {
-            codeLists[codeListOid].sources.itemDefs.forEach( itemDefOid => {
+            codeLists[codeListOid].sources.itemDefs.forEach(itemDefOid => {
                 itemDefOids.push(itemDefOid);
             });
         });
@@ -418,7 +418,17 @@ class ConnectedCodeListTable extends React.Component {
             codeListOids,
             itemDefOids,
         };
-        this.props.deleteCodeLists(deleteObj);
+        // check if the prompt option is enabled and codelist being deleted are used by variables
+        if (this.props.showDeleteCodeListWarning && itemDefOids.length !== 0) {
+            // if the check is enabled and codelists are used by some variables, open modal to confirm deletion
+            this.props.openModal({
+                type: 'DELETE_CODELISTS',
+                props: { deleteObj }
+            });
+        } else {
+            // otherwise, delete the codelists straightaway
+            this.props.deleteCodeLists(deleteObj);
+        }
         this.setState({ selectedRows: [] });
     }
 
@@ -433,10 +443,10 @@ class ConnectedCodeListTable extends React.Component {
         } else {
             // If the item is going to be removed;
             if (selectedRows.includes(row.oid)) {
-                selectedRows.splice(selectedRows.indexOf(row.oid),1);
+                selectedRows.splice(selectedRows.indexOf(row.oid), 1);
             }
         }
-        this.setState({selectedRows});
+        this.setState({ selectedRows });
         return true;
     }
 
@@ -445,11 +455,11 @@ class ConnectedCodeListTable extends React.Component {
         if (isSelected === true) {
             // If all rows are going to be selected;
             selectedRows = rows
-                .map( row => (row.oid));
+                .map(row => (row.oid));
         } else {
             selectedRows = [];
         }
-        this.setState({selectedRows});
+        this.setState({ selectedRows });
         return true;
     }
 
@@ -460,15 +470,15 @@ class ConnectedCodeListTable extends React.Component {
         this.props.codeListOrder.forEach((codeListOid, index) => {
             const originCL = codeListsRaw[codeListOid];
             let currentCL = {
-                oid            : originCL.oid,
-                name           : originCL.name,
-                dataType       : originCL.dataType,
-                codeListType   : originCL.codeListType,
-                formatName     : originCL.formatName,
-                orderNumber    : originCL.orderNumber,
-                linkedCodeList : originCL.linkedCodeListOid !== undefined ? codeListsRaw[originCL.linkedCodeListOid].name : undefined,
-                defineVersion  : this.props.defineVersion,
-                stdConstants   : this.props.stdConstants,
+                oid: originCL.oid,
+                name: originCL.name,
+                dataType: originCL.dataType,
+                codeListType: originCL.codeListType,
+                formatName: originCL.formatName,
+                orderNumber: originCL.orderNumber,
+                linkedCodeList: originCL.linkedCodeListOid !== undefined ? codeListsRaw[originCL.linkedCodeListOid].name : undefined,
+                defineVersion: this.props.defineVersion,
+                stdConstants: this.props.stdConstants,
             };
             let sources = getSourceLabels(originCL.sources, this.props.mdv);
             if (sources.hasOwnProperty('itemDefs')) {
@@ -478,9 +488,9 @@ class ConnectedCodeListTable extends React.Component {
             }
             if (originCL.codeListType !== 'external') {
                 currentCL.standardData = {
-                    alias                : originCL.alias,
-                    standardOid          : originCL.standardOid,
-                    cdiscSubmissionValue : originCL.cdiscSubmissionValue,
+                    alias: originCL.alias,
+                    standardOid: originCL.standardOid,
+                    cdiscSubmissionValue: originCL.cdiscSubmissionValue,
                 };
             } else {
                 currentCL.standardData = { ...originCL.externalCodeList };
@@ -496,29 +506,29 @@ class ConnectedCodeListTable extends React.Component {
 
         // Editor settings
         const cellEditProp = {
-            mode           : 'dbclick',
-            blurToSave     : true,
-            beforeSaveCell : this.onBeforeSaveCell
+            mode: 'dbclick',
+            blurToSave: true,
+            beforeSaveCell: this.onBeforeSaveCell
         };
 
         let selectRowProp;
         if (this.props.showRowSelect) {
             selectRowProp = {
-                mode          : 'checkbox',
-                clickToSelect : true,
-                onSelect      : this.onRowSelected,
-                onSelectAll   : this.onAllRowSelected,
-                columnWidth   : '48px',
+                mode: 'checkbox',
+                clickToSelect: true,
+                onSelect: this.onRowSelected,
+                onSelectAll: this.onAllRowSelected,
+                columnWidth: '48px',
+                selected: this.state.selectedRows,
             };
         } else {
             selectRowProp = undefined;
         }
 
         const options = {
-            toolBar  : this.createCustomToolBar,
-            btnGroup : this.createCustomButtonGroup
+            toolBar: this.createCustomToolBar,
+            btnGroup: this.createCustomButtonGroup
         };
-
 
         return (
             <React.Fragment>
@@ -531,8 +541,8 @@ class ConnectedCodeListTable extends React.Component {
                     remote={ true }
                     version='4'
                     cellEdit={this.props.reviewMode || this.props.showRowSelect ? undefined : cellEditProp}
-                    keyBoardNav={this.props.showRowSelect ? false : {enterToEdit: true}}
-                    headerStyle={{backgroundColor: indigo[500], color: grey[200], fontSize: '16px'}}
+                    keyBoardNav={this.props.showRowSelect ? false : { enterToEdit: true }}
+                    headerStyle={{ backgroundColor: indigo[500], color: grey[200], fontSize: '16px' }}
                     selectRow={selectRowProp}
                 >
                     {renderColumns(this.state.columns)}
@@ -555,13 +565,13 @@ class ConnectedCodeListTable extends React.Component {
 }
 
 ConnectedCodeListTable.propTypes = {
-    codeLists     : PropTypes.object.isRequired,
-    stdCodeLists  : PropTypes.object.isRequired,
-    stdConstants  : PropTypes.object.isRequired,
-    mdv           : PropTypes.object.isRequired,
-    classes       : PropTypes.object.isRequired,
-    defineVersion : PropTypes.string.isRequired,
-    reviewMode    : PropTypes.bool,
+    codeLists: PropTypes.object.isRequired,
+    stdCodeLists: PropTypes.object.isRequired,
+    stdConstants: PropTypes.object.isRequired,
+    mdv: PropTypes.object.isRequired,
+    classes: PropTypes.object.isRequired,
+    defineVersion: PropTypes.string.isRequired,
+    reviewMode: PropTypes.bool,
 };
 
 const CodeListTable = connect(mapStateToProps, mapDispatchToProps)(ConnectedCodeListTable);
